stock-predictor-app/api/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
yfinance==0.2.28
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
fastapi-utils==0.2.1
websockets==11.0.3
plotly
textblob
finnhub-python
statsmodels
prophet
tensorflow
newsapi-python 